Rename package from com.qompium.fibricheckcamerasdk to com.qompium.fibricheck.camerasdk
