+++ README.md
@@ -61,7 +61,7 @@ dependencies {
 Once the dependency is correctly added, the SDK is available in your code.
 
 ```java
-import com.qompium.fibricheck_camera_sdk.*;
+import com.qompium.fibricheck.camerasdk.*;
 ```
 
 ## License
