+++ example/src/main/java/com/qompium/fibricheckexample/MainActivity.java
@@ -18,7 +18,7 @@ import com.qompium.fibricheckexample.databinding.ActivityMainBinding;
 import android.view.Menu;
 import android.view.MenuItem;
 
-import com.qompium.fibricheck_camera_sdk.FibriChecker;
+import com.qompium.fibricheck.camerasdk.FibriChecker;
 
 public class MainActivity extends AppCompatActivity {
