+++ fibricheck-camera-sdk/src/main/java/com/qompium/fibricheck/camerasdk/listeners/SensorListener.java
@@ -1,4 +1,4 @@
-package com.qompium.fibricheck_camera_sdk.listeners;
+package com.qompium.fibricheck.camerasdk.listeners;
 
 import android.content.Context;
 import android.hardware.Sensor;
